eureka-manim/requirements.txt
manim>=0.18.0
opencv-python>=4.9.0
pysrt>=1.1.2
azure-cognitiveservices-speech>=1.40.0
python-dotenv>=1.0.0
